backend/requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
greenlet==3.1.1
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==3.0.1
mysql-connector-python==9.1.0
mysqlclient==2.2.4
pymysql==1.1.1
typing-extensions==4.12.2
werkzeug==3.0.4
